+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### 0.0.9
 - Added ability to fix broken require statements.
 - Added ability to manually fix the import/require statements when more than one path is matched.
+- Added support of JavaScript, TypeScript and their React expansions.
 
 ### 0.0.8
 - Added ability to read code template from a JavaScript file.